Verbal and physical aggression scores from Singh et al. (2007)

This is a dataset originally described in Singh et al. (2007), and digitized by Rumen Manolov using plot digitizer software and used to illustrate a number of single case design analysis approaches in Manolov & Moeyaert (2016). It is also used by Verboon & Peters (2017) to illustrate the piecewiseRegr and the genlog functions.

data("Singh")

Format

A data frame with 56 observations on the following 6 variables.

tier

A numeric subject identifier.

id

A character subject identifier (i.e. a name).

time

An index of the measurement moment.

phase

A dummy variable indicating the phase of the experiment: 0 means that treatment has not yet started, 1 means that treatment has started.

score_physical

The subjects' scores on physical aggression.

score_verbal

The subjects' scores on verbal aggression.

Source

See Rumen Manolov's Open Science Framework repository at https://osf.io/t6ws6 for the tutorial and the original dataset.

References

Singh, N. N., Lancioni, G. E., Winton, A. S., Adkins, A. D., Wahler, R. G., Sabaawi, M., & Singh, J. (2007). Individuals with mental illness can control their aggressive behavior through mindfulness training. Behavior Modification, 31(3), 313-328. https://doi.org/10.1177/0145445506293585

Manolov, R., & Moeyaert, M. (2017). How Can Single-Case Data Be Analyzed? Software Resources, Tutorial, and Reflections on Analysis. Behavior Modification, 41(2), 179-228. https://doi.org/10.1177/0145445516664307

Verboon, P. & Peters, G.-J. Y. (2017) Applying the generalised logistic model in SCD to deal with ceiling effects. PsyArXiv http://INSERTLINK

See also

Examples

### To load the data, use:
data(Singh);