exceptionalScores

A function to detect participants that consistently respond exceptionally.

exceptionalScores(dat, items = NULL, exception = 0.025, totalOnly = TRUE,
                  append = TRUE, both = TRUE, silent = FALSE,
                  suffix = "_isExceptional", totalVarName = "exceptionalScores")

Arguments

dat
The dataframe containing the variables to inspect, or the vector to inspect (but for vectors, exceptionalScore might be more useful).

items
The names of the variables to inspect.

exception
When an item will be considered exceptional, passed on as prob to exceptionalScore.

totalOnly
Whether to return only the number of exceptional scores for each row in the dataframe, or for each inspected item, which values are exceptional.

append
Whether to return the supplied dataframe with the new variable(s) appended (if TRUE), or whether to only return the new variable(s) (if FALSE).

both
Whether to look for both low and high exceptional scores (TRUE) or not (FALSE; see exceptionalScore).

silent
Can be used to suppress messages.

suffix
If not returning the total number of exceptional values, for each inspected variable, a new variable is returned indicating which values are exceptional. The text string is appended to each original variable name to create the new variable names.

totalVarName
If returning only the total number of exceptional values, and appending these to the provided dataset, this text string is used as variable name.

Value

Either a vector containing the number of exceptional values, a dataset containing, for each inspected variable, which values are exceptional, or the provided dataset where either the total or the exceptional values for each variable are appended.

See also

Examples

exceptionalScores(mtcars)

#> No items specified: extracting all variable names in dataframe.

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>                     exceptionalScores
#> Mazda RX4                           0
#> Mazda RX4 Wag                       0
#> Datsun 710                          0
#> Hornet 4 Drive                      0
#> Hornet Sportabout                   0
#> Valiant                             0
#> Duster 360                          0
#> Merc 240D                           0
#> Merc 230                            1
#> Merc 280                            0
#> Merc 280C                           0
#> Merc 450SE                          0
#> Merc 450SL                          0
#> Merc 450SLC                         0
#> Cadillac Fleetwood                  1
#> Lincoln Continental                 1
#> Chrysler Imperial                   0
#> Fiat 128                            0
#> Honda Civic                         2
#> Toyota Corolla                      2
#> Toyota Corona                       0
#> Dodge Challenger                    0
#> AMC Javelin                         0
#> Camaro Z28                          0
#> Pontiac Firebird                    0
#> Fiat X1-9                           0
#> Porsche 914-2                       0
#> Lotus Europa                        1
#> Ford Pantera L                      1
#> Ferrari Dino                        0
#> Maserati Bora                       2
#> Volvo 142E                          0