Frequency tables

Function to show frequencies in a manner similar to what SPSS' "FREQUENCIES" command does. Note that frequency is an alias for freq.

freq(vector, digits = 1, nsmall=1, transposed=FALSE,
     round=1, plot=FALSE, plotTheme = theme_bw())
frequencies(..., digits = 1, nsmall = 1,
            transposed = FALSE, round = 1,
            plot = FALSE, plotTheme = theme_bw())

Arguments

vector
A vector of values to compute frequencies for.

digits
Minimum number of significant digits to show in result.

nsmall
Minimum number of digits after the decimal point to show in the result.

transposed
Whether to transpose the results when printing them (this can be useful for blind users).

round
Number of digits to round the results to (can be used in conjunction with digits to determine format of results).

plot
If true, a histogram is shown of the variable.

plotTheme
The ggplot2 theme to use.

…
The variables of which to provide frequencies

Value

An object with several elements, the most notable of which is:

dat

A dataframe with the frequencies

For frequencies, these objects are in a list of their own.
Examples

### Create factor vector
ourFactor <- factor(mtcars$gear, levels=c(3,4,5),
                    labels=c("three", "four", "five"));
### Add some missing values
factorWithMissings <- ourFactor;
factorWithMissings[10] <- factorWithMissings[20] <- NA;

### Show frequencies
freq(ourFactor);

#>             Frequencies Perc.Total Perc.Valid Cumulative
#> three                15       46.9       46.9       46.9
#> four                 12       37.5       37.5       84.4
#> five                  5       15.6       15.6      100.0
#> Total valid          32      100.0      100.0           

freq(factorWithMissings);

#>              Frequencies Perc.Total Perc.Valid Cumulative
#> three                 15       46.9       50.0       50.0
#> four                  10       31.2       33.3       83.3
#> five                   5       15.6       16.7      100.0
#> Total valid           30       93.8      100.0           
#> NA (missing)           2        6.2                      
#> Total                 32      100.0                      

### ... Or for all of them at one
frequencies(ourFactor, factorWithMissings);

#> ### Frequencies for 'ourFactor'
#> 
#>             Frequencies Perc.Total Perc.Valid Cumulative
#> three                15       46.9       46.9       46.9
#> four                 12       37.5       37.5       84.4
#> five                  5       15.6       15.6      100.0
#> Total valid          32      100.0      100.0           
#> 
#> ### Frequencies for 'factorWithMissings'
#> 
#>              Frequencies Perc.Total Perc.Valid Cumulative
#> three                 15       46.9       50.0       50.0
#> four                  10       31.2       33.3       83.3
#> five                   5       15.6       16.7      100.0
#> Total valid           30       93.8      100.0           
#> NA (missing)           2        6.2                      
#> Total                 32      100.0                      
#>