The distribution of Omega Squared

These functions use some conversion to and from the F distribution to provide the Omega Squared distribution.

domegaSq(x, df1, df2, populationOmegaSq = 0)
pomegaSq(q, df1, df2, populationOmegaSq = 0, lower.tail = TRUE)
qomegaSq(p, df1, df2, populationOmegaSq = 0, lower.tail = TRUE)
romegaSq(n, df1, df2, populationOmegaSq = 0)

Arguments

x, q
Vector of quantiles, or, in other words, the value(s) of Omega Squared.

p
Vector of probabilites (p-values).

df1, df2
Degrees of freedom for the numerator and the denominator, respectively.

n
Desired number of Omega Squared values.

populationOmegaSq
The value of Omega Squared in the population; this determines the center of the Omega Squared distribution. This has not been implemented yet in this version of userfriendlyscience. If anybody has the inverse of convert.ncf.to.omegasq for me, I'll happily integrate this.

lower.tail
logical; if TRUE (default), probabilities are the likelihood of finding an Omega Squared smaller than the specified value; otherwise, the likelihood of finding an Omega Squared larger than the specified value.

Details

The functions use convert.omegasq.to.f and convert.f.to.omegasq to provide the Omega Squared distribution.

Value

domegaSq gives the density, pomegaSq gives the distribution function, qomegaSq gives the quantile function, and romegaSq generates random deviates.

See also

Examples

### Generate 10 random Omega Squared values
romegaSq(10, 66, 3);

#>  [1]  0.64442335  0.06071611 -0.29161453 -0.33789699  0.09548003  0.69827477
#>  [7]  0.83047431  0.46244299  0.41944212  0.93617402

### Probability of findings an Omega Squared
### value smaller than .06 if it's 0 in the population
pomegaSq(.06, 66, 3);

#> [1] 0.4280867