userfriendlyscience methods to pander objects

These methods try to provide output that's ready for R Markdown. Note that they are not all documented; most of them are quite straightforward.

# S3 method for freq
pander(x, ...)
# S3 method for meanDiff
pander(x, digits=x$digits, powerDigits=x$digits + 2, ...)
# S3 method for normalityAssessment
pander(x, headerPrefix = "#####", suppressPlot = FALSE, ...)
# S3 method for dataShape
pander(x, digits=x$input$digits, extraNotification=TRUE, ...)
# S3 method for associationMatrix
pander(x, info = x$input$info, file = x$input$file, ...)
# S3 method for crossTab
pander(x, digits = x$input$digits,
                          pValueDigits=x$input$pValueDigits, ...)
# S3 method for oneway
pander(x, digits = x$input$digits,
                        pvalueDigits=x$input$pvalueDigits,
                        headerStyle = "**",
                        na.print="", ...)
# S3 method for regr
pander(x, digits=x$input$digits,
                      pvalueDigits=x$input$pvalueDigits, ...)
# S3 method for descr
pander(x, headerPrefix = "", headerStyle = "**", ...)
# S3 method for examine
pander(x, headerPrefix = "", headerStyle = "**",
                         secondaryHeaderPrefix = "", secondaryHeaderStyle =
                         "*", ...)
# S3 method for examineBy
pander(x, headerPrefix = "", headerStyle = "**",
                           secondaryHeaderPrefix = "", secondaryHeaderStyle =
                           "*", tertairyHeaderPrefix = "--> ",
                           tertairyHeaderStyle = "", separator = paste0("\n\n",
                           repStr("-", 10), "\n\n"), ...)
# S3 method for frequencies
pander(x, prefix = "###", ...)

Arguments

x
The object to print.

digits
The number of significant digits to print.

powerDigits
Number of digits to use when printing the power.

headerPrefix, secondaryHeaderPrefix, tertairyHeaderPrefix, prefix
Prefix for headers, can be used to output headers for pandoc using R Markdown by specifying e.g. '####' for a level 4 header.

headerStyle, secondaryHeaderStyle, tertairyHeaderStyle
A character value to pre- and append to the header. This can be used to make the header appear bold ('**') or italic ('*') when not using an actual header (see headerPrefix).

separator
Separator to show between sections of output.

suppressPlot
Whether to suppress printing plots.

pValueDigits
Output to produce; see /code/linkrMatrix for details.

info, file
Output to produce and file to write to; see /code/linkassociationMatrix for details.

extraNotification
Whether an extra notification about the type of skewness and kurtosis returned by dataShape is shown.

pvalueDigits
The number of digits to show for p-values; smaller p-values will be shown as <.001 or <.0001 etc.

na.print
What to print for missing values, for example for a oneway anova table.

...
Additional arguments that are passed on to the print functions when it is called.

Value

These printing methods use cat, cat0, and grid.draw to print stuff.

Examples

userfriendlyscience:::pander.oneway(oneway(y=ChickWeight$weight, x=ChickWeight$Diet));

#> 
#> 
#> **Oneway Anova for y=weight and x=Diet (groups: 1, 2, 3 & 4)**
#> 
#> Omega squared: 95% CI = [.02; .09], point estimate = .05  
#> Eta Squared: 95% CI = [.02; .08], point estimate = .05
#> 
#> 
#> ------------------------------------------------------------------------------
#>              &nbsp;                   SS       Df       MS        F       p   
#> -------------------------------- ------------ ----- ---------- ------- -------
#>    **Between groups (error +      155862.66     3    51954.22   10.81   <.001 
#>            effect)**                                                          
#> 
#>  **Within groups (error only)**   2758693.27   574   4806.09                  
#> ------------------------------------------------------------------------------
#> 
#> 
#>