removeExceptionalValues

A function to replace exceptional values with NA. This can be used to quickly remove impossible values, for example, when participants entered their age as 344.

removeExceptionalValues(dat, items = NULL, exception = 0.005,
                        silent = FALSE, stringsAsFactors = FALSE)

Arguments

dat
The dataframe containing the items to inspect.

items
The items to inspect.

exception
How rare a value must be to be considered exceptional (and replaced by NA).

silent
Can be used to suppress messages.

stringsAsFactors
Whether to convert strings to factors when creating a dataframe from lapply output.

Details

Note that exceptional values may be errors (e.g. participants accidently pressed a key twice, or during data entry, something went wrong), but they may also be indicative of participants who did not seriously participate in the study. Therefore, it is advised to first use exceptionalScores to look for patterns where participants enter many exceptional scores.

Value

The dataframe, with exceptional values replaced by NA.

See also

Examples

removeExceptionalValues(mtcars, exception=.1);

#> No items specified: extracting all variable names in dataframe.

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6  18.1   6 225.0 105   NA 3.460    NA  1  0    3    1
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8  24.4   4 146.7  NA 3.69 3.190 20.00  1  0    4    2
#> 9  22.8   4 140.8  95 3.92 3.150    NA  1  0    4    2
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15   NA   8    NA 205   NA    NA 17.98  0  0    3    4
#> 16   NA   8    NA 215 3.00    NA 17.82  0  0    3    4
#> 17 14.7   8    NA 230 3.23    NA 17.42  0  0    3    4
#> 18   NA   4    NA  66 4.08 2.200 19.47  1  1    4    1
#> 19 30.4   4    NA  NA   NA    NA 18.52  1  1    4    2
#> 20   NA   4    NA  NA 4.22    NA 19.90  1  1    4    1
#> 21 21.5   4 120.1  97 3.70 2.465    NA  1  0    3    1
#> 22 15.5   8 318.0 150   NA 3.520 16.87  0  0    3    2
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24   NA   8 350.0 245 3.73 3.840    NA  0  0    3    4
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 27 26.0   4 120.3  91   NA 2.140 16.70  0  1    5    2
#> 28 30.4   4  95.1 113 3.77    NA 16.90  1  1    5    2
#> 29 15.8   8 351.0  NA 4.22 3.170    NA  0  1    5    4
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5   NA
#> 31 15.0   8 301.0  NA 3.54 3.570    NA  0  1    5   NA
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2