testRetestReliability

The testRetestReliability function is a convenient interface to testRetestAlpha and testRetestCES.

testRetestReliability(dat = NULL, moments = NULL,
                      testDat = NULL, retestDat = NULL,
                      parallelTests = 'means',
                      sortItems = FALSE, convertToNumeric = TRUE,
                      digits=2)

Arguments

dat
A dataframe. This dataframe must contain the items in the scale at both measurement moments. If no dataframe is specified, a dialogue will be launched to allow the user to select an SPSS datafile. If only one dataframe is specified, either the items have to be ordered chronologically (i.e. first all items for the first measurement, then all items for the second measurement), or the vector 'moments' has to be used to indicate, for each item, to which measurement moment it belongs. The number of columns in this dataframe MUST be even! Note that instead of providing this dataframe, the items of each measurement moment can be provided separately in testDat and retestDat as well.

moments
Used to indicate to which measurement moment each item in 'dat' belongs; should be a vector with the same length as dat has columns, and with two possible values (e.g. 1 and 2).

testDat, retestDat
Dataframes with the items for each measurement moment: note that the items have to be in the same order (unless sortItems is TRUE).

parallelTests
A vector indicating which items belong to which parallel test; like the moments vector, this should have two possible values (e.g. 1 and 2). Alternatively, it can be character value with 'means' or 'variances'; in this case, parallelSubscales will be used to create roughly parallel halves.

sortItems
If true, the columns (items) in each dataframe are ordered alphabetically before starting. This can be convenient to ensure that the order of the items at each measurement moment is the same.

convertToNumeric
When TRUE, the function will attempt to convert all vectors in the dataframes to numeric.

digits
Number of digits to show when printing the output

Details

This function calls both testRetestAlpha and testRetestCES to compute and print measures of the test-retest reliability.

Value

An object with the input and several output variables. Most notably:

input

Input specified when calling the function

intermediate

Intermediate values and objects computed to get to the final results

output$testRetestAlpha

The value of the test-retest alpha coefficient.

output$testRetestCES

The value of the test-retest Coefficient of Equivalence and Stability.

Examples

# NOT RUN {
### This will prompt the user to select an SPSS file
testRetestReliability();
# }

### Load data from simulated dataset testRetestSimData (which
### satisfies essential tau-equivalence).
data(testRetestSimData);

### The first column is the true score, so it's excluded in this example.
exampleData <- testRetestSimData[, 2:ncol(testRetestSimData)];

### Compute test-retest alpha coefficient
testRetestReliability(exampleData);

#>                          Items at time 1: t0_item1, t0_item2, t0_item3, t0_item4, t0_item5, t0_item6, t0_item7, t0_item8, t0_item9, t0_item10
#>                          Items at time 2: t1_item1, t1_item2, t1_item3, t1_item4, t1_item5, t1_item6, t1_item7, t1_item8, t1_item9, t1_item10
#>                             Observations: 250
#>            Test-retest Alpha Coefficient: 0.44
#> Coefficient of Equivalence and Stability: 0.43
#> 
#> To help assess whether the subscales (automatically generated using means) are parallel, here are the means and variances:
#> Mean subscale a1, time 1: 109.29 (variance = 219.59)
#> Mean subscale a2, time 1: 167.77 (variance = 556.03)
#> Mean subscale a1, time 2: 110.9 (variance = 189.55)
#> Mean subscale a2, time 2: 170.85 (variance = 613.37)